refactor(stations):Refactor stations into multiple components
